refactor(adr-102): split serialization.py into a serialization/ package (P6d)
The 1573-line api/lib/serialization.py held four distinct responsibilities and
was flagged repeatedly by the >800-line quality gate. Split at the natural class
seams into a package with zero caller changes:

- serialization/format.py   — BackupFormat + KgBackupV2Reader + KG_BACKUP_FORMAT_VERSION
                              (schema probe + pure read path; 196 lines)
- serialization/exporter.py — DataExporter + _parse_nullable_int (export path; 703 lines)
- serialization/importer.py — DataImporter + _execute_with_age_retry/_progress/_run_parallel
                              + epoch primitives (clone/merge writer; 698 lines)
- serialization/__init__.py — re-exports the public surface (BackupFormat, KgBackupV2Reader,
                              DataExporter, DataImporter, KG_BACKUP_FORMAT_VERSION)

All existing 'from api.lib.serialization import ...' callers (8 modules + tests)
keep working unchanged via the __init__ re-export. No cross-class references
between the three modules, so no circular imports.

Incidental cleanup: dropped the vestigial 'sys.path.insert(...)' hack and the
now-unused sys/Path/Colors imports — the absolute 'from api.app.lib.age_client
import AGEClient' resolves via the real PYTHONPATH, not the hack (which inserted
the wrong dir anyway). Behavior-preserving: full ADR-102 + backup test slice
(93 passed, 1 skipped) green; app import + public-surface assertions pass.

exporter/importer remain ~700 lines each but are single cohesive classes — the
three-class seam is the natural split; no further non-artificial seam exists.
